services/bq-writer/requirements.txt
Flask==3.*
gunicorn==22.*
google-cloud-bigquery==3.*
google-auth==2.*
